+++ base/exports.jl
@@ -778,6 +778,7 @@ export
 # misc
     atexit,
     atreplinit,
+    activate_module,
     exit,
     ntuple,